fix number type MonthlyChargesTranscribing, change int to float32

+++ infrastructure/bunny_request_rest/create_video_library_response.go
@@ -83,7 +83,7 @@ type CreateVideoLibraryResponse struct {
 	DrmCostPerLicenseOverride               *int             `json:"DrmCostPerLicenseOverride"`
 	TranscribingPriceOverride               *int             `json:"TranscribingPriceOverride"`
 	PremiumEncodingPriceOverride            *int             `json:"PremiumEncodingPriceOverride"`
-	MonthlyChargesTranscribing              int              `json:"MonthlyChargesTranscribing"`
+	MonthlyChargesTranscribing              float32          `json:"MonthlyChargesTranscribing"`
 	MonthlyChargesPremiumEncoding           int              `json:"MonthlyChargesPremiumEncoding"`
 	MonthlyChargesEnterpriseDrm             int              `json:"MonthlyChargesEnterpriseDrm"`
 }
